REFACTORIZED the Admin panel so that the index is shorter
