+++ frontend/src/components/AdminPanel/index.jsx
@@ -1,21 +1,8 @@
 import React, { useEffect, useState, useRef } from "react";
-import { ToastContainer, toast } from "react-toastify";
+import { ToastContainer } from "react-toastify";
 import "react-toastify/dist/ReactToastify.css";
 
 import TabNavigation from "./subcomponents/TabNavigation";
-import ColumnSelector from "./subcomponents/ColumnSelector";
-import Pagination from "./subcomponents/Pagination";
-import {
-  UsuarioTable,
-  InvestigadorTable,
-  ProyectoTable,
-} from "./subcomponents/Tables/tables";
-import {
-  UsuarioCards,
-  InvestigadorCards,
-  ProyectoCards,
-} from "./subcomponents/Cards/cards";
-
 import {
   UsuarioForm,
   InvestigadorForm,
@@ -23,11 +10,23 @@ import {
   DeleteConfirmation,
 } from "./subcomponents/Forms/forms";
 
+// Custom hooks
 import { useAdminPanel } from "./hooks/useAdminPanel";
-import { useTableControls, columnOrders } from "./hooks/useTableControls";
+import { useTableControls } from "./hooks/useTableControls";
+
+// Services
 import { usuarioService } from "../../api/services/usuarioService";
 import { investigadorService } from "../../api/services/investigadorService";
 import { proyectoService } from "../../api/services/proyectoService";
+
+// Utils and components
+import { showNotification, copyToClipboard } from "./utils/notificationsUtils";
+import { setupAdminPanelStyles } from "./styles/adminPanelStyles";
+import { getTabData } from "./utils/dataUtils";
+import CreateButton from "./components/Buttons/CreateButton";
+import ViewControls from "./components/Controls/ViewControls";
+import ContentDisplay from "./components/Content/ContentDisplay";
+import LoadingSpinner from "./components/Loaders/LoadingSpinner";
 
 function AdminPanel() {
   const [contentReady, setContentReady] = useState(false);
@@ -70,49 +69,11 @@ function AdminPanel() {
     setColumnsDropdownOpen,
     contentRef,
     isMobile,
-    // eslint-disable-next-line no-unused-vars
-    getOrderedVisibleColumns,
   } = useTableControls();
 
   const columnToggleRef = useRef(null);
 
-  const showNotification = (message) => {
-    toast(message, {
-      position: "bottom-right",
-      autoClose: 3000, 
-      hideProgressBar: false,
-      closeOnClick: false,
-      pauseOnHover: true,
-      draggable: false, 
-      closeButton: false, 
-      progress: undefined,
-      theme: "dark",
-      style: {
-        background: "linear-gradient(to right, #1e40af, #3b82f6)",
-        color: "white",
-        borderRadius: "8px",
-        boxShadow: "0 4px 12px rgba(59, 130, 246, 0.3)",
-        fontSize: "16px",
-        padding: "12px 16px",
-        minHeight: "40px",
-        display: "flex",
-        alignItems: "center",
-      },
-      bodyStyle: {
-        fontFamily: "inherit",
-        fontSize: "16px",
-        fontWeight: "500",
-      },
-    });
-  };
-
-  const copyToClipboard = (text) => {
-    navigator.clipboard
-      .writeText(text)
-      .then(() => showNotification("Copiado al portapapeles"))
-      .catch(() => showNotification("Error al copiar"));
-  };
-
+  // Setup content ready effect
   useEffect(() => {
     let timer;
     if (!loading) {
@@ -123,75 +84,9 @@ function AdminPanel() {
     return () => clearTimeout(timer);
   }, [loading]);
 
+  // Setup styles
   useEffect(() => {
-    const styleEl = document.createElement("style");
-    styleEl.id = "admin-panel-styles";
-
-    styleEl.innerHTML = `
-      @keyframes fadeIn {
-        from { opacity: 0; transform: translateY(10px); }
-        to { opacity: 1; transform: translateY(0); }
-      }
-      
-      @keyframes pulse {
-        0% { box-shadow: 0 0 0 0 rgba(59, 130, 246, 0.5); }
-        70% { box-shadow: 0 0 0 10px rgba(59, 130, 246, 0); }
-        100% { box-shadow: 0 0 0 0 rgba(59, 130, 246, 0); }
-      }
-      
-      .admin-fadeIn { animation: fadeIn 0.5s ease-out forwards; opacity: 0; }
-      .no-scrollbar { overflow: hidden; }
-      .admin-pulse { animation: pulse 2s infinite; }
-      .view-transition { transition: opacity 0.3s ease-out, transform 0.3s ease-out; }
-      
-      #column-dropdown { position: absolute; z-index: 50; }
-      
-      .notification {
-        position: fixed;
-        bottom: 20px;
-        right: 20px;
-        background: rgba(59, 130, 246, 0.9);
-        color: white;
-        padding: 10px 20px;
-        border-radius: 4px;
-        transform: translateY(100px);
-        opacity: 0;
-        transition: all 0.3s ease-out;
-        z-index: 9999;
-      }
-      
-      .notification.show { transform: translateY(0); opacity: 1; }
-      
-      .content-container {
-        visibility: hidden;
-        opacity: 0;
-        transition: visibility 0.3s, opacity 0.3s;
-      }
-      
-      .content-container.ready { visibility: visible; opacity: 1; }
-      .pagination-container { min-height: 40px; }
-
-      /* Estilos específicos para pantallas pequeñas */
-      @media (max-width: 639px) {
-        .mobile-controls-grid {
-          display: grid;
-          grid-template-columns: 1fr 1fr;
-          gap: 8px;
-        }
-        
-        .control-button {
-          width: 100%;
-          justify-content: center;
-        }
-        
-        .mobile-full-width {
-          grid-column: 1 / -1;
-          width: 100%;
-        }
-      }
-    `;
-
-    document.head.appendChild(styleEl);
+    const styleEl = setupAdminPanelStyles();
     return () => {
       if (document.getElementById("admin-panel-styles")) {
         document.head.removeChild(styleEl);
@@ -199,6 +94,7 @@ function AdminPanel() {
     };
   }, []);
 
+  // CRUD Handlers
   const handleCreate = (type) => {
     setFormModal({
       isOpen: true,
@@ -279,52 +175,7 @@ function AdminPanel() {
     });
   };
 
-  const renderCreateButton = () => {
-    let type = "";
-    let label = "";
-
-    switch (activeTab) {
-      case "usuarios":
-        type = "usuario";
-        label = "Usuario";
-        break;
-      case "investigadores":
-        type = "investigador";
-        label = "Investigador";
-        break;
-      case "proyectos":
-        type = "proyecto";
-        label = "Proyecto";
-        break;
-      default:
-        type = "usuario";
-        label = "Usuario";
-    }
-
-    return (
-      <button
-        className="cursor-pointer flex items-center gap-2 px-4 py-2 bg-blue-600 hover:bg-blue-700 text-white rounded-md transition-colors w-full sm:w-auto justify-center sm:justify-start"
-        onClick={() => handleCreate(type)}
-      >
-        <svg
-          xmlns="http://www.w3.org/2000/svg"
-          className="h-5 w-5"
-          fill="none"
-          viewBox="0 0 24 24"
-          stroke="currentColor"
-        >
-          <path
-            strokeLinecap="round"
-            strokeLinejoin="round"
-            strokeWidth={2}
-            d="M12 4v16m8-8H4"
-          />
-        </svg>
-        Crear {label}
-      </button>
-    );
-  };
-
+  // Render form based on type
   const renderForm = () => {
     const { type, item, isOpen } = formModal;
 
@@ -363,82 +214,7 @@ function AdminPanel() {
     }
   };
 
-  const getTabData = () => {
-    let type = "";
-
-    switch (activeTab) {
-      case "usuarios":
-        type = "usuario";
-        break;
-      case "investigadores":
-        type = "investigador";
-        break;
-      case "proyectos":
-        type = "proyecto";
-        break;
-      default:
-        type = "usuario";
-    }
-
-    // Obtener columnas visibles ordenadas según el orden definido
-    const orderedColumns = {
-      usuarios: sortColumnsByOrder(visibleColumns.usuarios || [], "usuarios"),
-      investigadores: sortColumnsByOrder(
-        visibleColumns.investigadores || [],
-        "investigadores"
-      ),
-      proyectos: sortColumnsByOrder(
-        visibleColumns.proyectos || [],
-        "proyectos"
-      ),
-    };
-
-    const tabConfig = {
-      usuarios: {
-        title: "Lista de Usuarios",
-        items: usuarios || [],
-        TableComponent: UsuarioTable,
-        CardComponent: UsuarioCards,
-        columns: orderedColumns.usuarios,
-        onEdit: (item) => handleEdit(type, item),
-        onDelete: (item) => handleDeleteClick(type, item),
-      },
-      investigadores: {
-        title: "Lista de Investigadores",
-        items: investigadores || [],
-        TableComponent: InvestigadorTable,
-        CardComponent: InvestigadorCards,
-        columns: orderedColumns.investigadores,
-        onEdit: (item) => handleEdit(type, item),
-        onDelete: (item) => handleDeleteClick(type, item),
-      },
-      proyectos: {
-        title: "Lista de Proyectos",
-        items: proyectos || [],
-        TableComponent: ProyectoTable,
-        CardComponent: ProyectoCards,
-        columns: orderedColumns.proyectos,
-        onEdit: (item) => handleEdit(type, item),
-        onDelete: (item) => handleDeleteClick(type, item),
-      },
-    };
-
-    return tabConfig[activeTab] || tabConfig.usuarios;
-  };
-
-  // Función auxiliar para ordenar columnas
-  const sortColumnsByOrder = (columns, tabName) => {
-    if (!columns || !Array.isArray(columns)) return [];
-
-    return [...columns].sort((a, b) => {
-      const orderA = columnOrders[tabName].indexOf(a);
-      const orderB = columnOrders[tabName].indexOf(b);
-      if (orderA === -1) return 1;
-      if (orderB === -1) return -1;
-      return orderA - orderB;
-    });
-  };
-
+  // Get data for current tab
   const {
     title,
     items,
@@ -447,7 +223,7 @@ function AdminPanel() {
     columns,
     onEdit,
     onDelete,
-  } = getTabData();
+  } = getTabData(activeTab, visibleColumns, usuarios, investigadores, proyectos, handleEdit, handleDeleteClick);
 
   return (
     <div className="mt-4 pb-10 w-full px-2 sm:px-4 overflow-x-hidden no-scrollbar">
@@ -461,15 +237,7 @@ function AdminPanel() {
       <TabNavigation activeTab={activeTab} changeTab={changeTab} />
 
       {loading ? (
-        <div className="flex justify-center items-center h-64">
-          <div className="w-16 h-16 relative">
-            <div className="w-16 h-16 rounded-full border-4 border-blue-500/20 border-t-blue-500 animate-spin"></div>
-            <div
-              className="w-16 h-16 rounded-full absolute top-0 left-0 border-4 border-transparent border-b-indigo-500 animate-spin"
-              style={{ animationDuration: "1.5s" }}
-            ></div>
-          </div>
-        </div>
+        <LoadingSpinner />
       ) : (
         <div className={`content-container ${contentReady ? "ready" : ""}`}>
           <div
@@ -483,205 +251,46 @@ function AdminPanel() {
               </h3>
 
               <div className="w-full sm:w-auto">
-                {renderCreateButton()}
+                <CreateButton 
+                  activeTab={activeTab} 
+                  onClick={handleCreate} 
+                />
               </div>
             </div>
 
-            {/* Controles de Visualización - Reorganizados para Móvil */}
-            <div className="mb-6">
-              {/* Vista para móviles (menor a 640px) */}
-              <div className="block sm:hidden mobile-controls-grid">
-                <button
-                  className={`flex items-center gap-2 px-3 py-2 rounded-md text-sm cursor-pointer transition-all duration-200 control-button ${
-                    viewMode === "table" && !isMobile
-                      ? "bg-gradient-to-r from-blue-600 to-indigo-600 text-white shadow-md"
-                      : "bg-gray-800/70 border border-blue-500/30 text-blue-300 hover:bg-gray-700/70 hover:border-blue-500/50"
-                  } ${isMobile ? "opacity-50 cursor-not-allowed" : ""}`}
-                  onClick={() => toggleViewMode("table")}
-                  disabled={isMobile}
-                >
-                  <svg
-                    xmlns="http://www.w3.org/2000/svg"
-                    width="16"
-                    height="16"
-                    fill="currentColor"
-                    viewBox="0 0 16 16"
-                  >
-                    <path d="M0 2a2 2 0 0 1 2-2h12a2 2 0 0 1 2 2v12a2 2 0 0 1-2 2H2a2 2 0 0 1-2-2V2zm15 2h-4v3h4V4zm0 4h-4v3h4V8zm0 4h-4v3h3a1 1 0 0 0 1-1v-2zm-5 3v-3H6v3h4zm-5 0v-3H1v2a1 1 0 0 0 1 1h3zm-4-4h4V8H1v3zm0-4h4V4H1v3zm5-3v3h4V4H6zm4 4H6v3h4V8z" />
-                  </svg>
-                  Tabla
-                </button>
-                <button
-                  className={`flex items-center gap-2 px-3 py-2 rounded-md text-sm cursor-pointer transition-all duration-200 control-button ${
-                    viewMode === "cards"
-                      ? "bg-gradient-to-r from-blue-600 to-indigo-600 text-white shadow-md"
-                      : "bg-gray-800/70 border border-blue-500/30 text-blue-300 hover:bg-gray-700/70 hover:border-blue-500/50"
-                  }`}
-                  onClick={() => toggleViewMode("cards")}
-                >
-                  <svg
-                    xmlns="http://www.w3.org/2000/svg"
-                    width="16"
-                    height="16"
-                    fill="currentColor"
-                    viewBox="0 0 16 16"
-                  >
-                    <path d="M0 4a2 2 0 0 1 2-2h12a2 2 0 0 1 2 2v8a2 2 0 0 1-2 2H2a2 2 0 0 1-2-2V4zm2-1a1 1 0 0 0-1 1v8a1 1 0 0 0 1 1h12a1 1 0 0 0 1-1V4a1 1 0 0 0-1-1H2z" />
-                    <path d="M4 5.5a.5.5 0 0 1 .5-.5h7a.5.5 0 0 1 .5.5v2a.5.5 0 0 1-.5.5h-7a.5.5 0 0 1-.5-.5v-2z" />
-                  </svg>
-                  Tarjetas
-                </button>
-                
-                {viewMode === "table" && !isMobile && (
-                  <div className="relative z-30" ref={columnToggleRef}>
-                    <ColumnSelector
-                      activeTab={activeTab}
-                      columnsDropdownOpen={columnsDropdownOpen}
-                      setColumnsDropdownOpen={setColumnsDropdownOpen}
-                      visibleColumns={visibleColumns}
-                      toggleColumn={toggleColumn}
-                      columnToggleRef={columnToggleRef}
-                    />
-                  </div>
-                )}
-                
-                {/* Selector de elementos por página - Ancho completo en móvil */}
-                <div className="flex justify-center items-center mobile-full-width mt-3">
-                  <div className="pt-2 flex items-center">
-                    <label className="text-sm text-gray-400 mr-2">Mostrar:</label>
-                    <select
-                      className="bg-gray-800 border border-gray-700 text-gray-300 rounded-md py-1 px-2 text-sm cursor-pointer transition-colors duration-200 hover:border-blue-500/30"
-                      value={itemsPerPage}
-                      onChange={(e) => {
-                        handleItemsPerPageChange(Number(e.target.value));
-                      }}
-                    >
-                      {[10, 15, 25, 50].map((value) => (
-                        <option key={value} value={value}>
-                          {value}
-                        </option>
-                      ))}
-                    </select>
-                  </div>
-                </div>
-              </div>
-              
-              {/* Vista para pantallas sm y mayores (>= 640px) */}
-              <div className="hidden sm:flex sm:flex-row justify-between items-center">
-                {/* Botones de Vista */}
-                <div className="flex items-center space-x-2 z-20">
-                  <button
-                    className={`flex items-center gap-2 px-3 py-2 rounded-md text-sm cursor-pointer transition-all duration-200 ${
-                      viewMode === "table" && !isMobile
-                        ? "bg-gradient-to-r from-blue-600 to-indigo-600 text-white shadow-md"
-                        : "bg-gray-800/70 border border-blue-500/30 text-blue-300 hover:bg-gray-700/70 hover:border-blue-500/50"
-                    } ${isMobile ? "opacity-50 cursor-not-allowed" : ""}`}
-                    onClick={() => toggleViewMode("table")}
-                    disabled={isMobile}
-                  >
-                    <svg
-                      xmlns="http://www.w3.org/2000/svg"
-                      width="16"
-                      height="16"
-                      fill="currentColor"
-                      viewBox="0 0 16 16"
-                    >
-                      <path d="M0 2a2 2 0 0 1 2-2h12a2 2 0 0 1 2 2v12a2 2 0 0 1-2 2H2a2 2 0 0 1-2-2V2zm15 2h-4v3h4V4zm0 4h-4v3h4V8zm0 4h-4v3h3a1 1 0 0 0 1-1v-2zm-5 3v-3H6v3h4zm-5 0v-3H1v2a1 1 0 0 0 1 1h3zm-4-4h4V8H1v3zm0-4h4V4H1v3zm5-3v3h4V4H6zm4 4H6v3h4V8z" />
-                    </svg>
-                    Tabla
-                  </button>
-                  <button
-                    className={`flex items-center gap-2 px-3 py-2 rounded-md text-sm cursor-pointer transition-all duration-200 ${
-                      viewMode === "cards"
-                        ? "bg-gradient-to-r from-blue-600 to-indigo-600 text-white shadow-md"
-                        : "bg-gray-800/70 border border-blue-500/30 text-blue-300 hover:bg-gray-700/70 hover:border-blue-500/50"
-                    }`}
-                    onClick={() => toggleViewMode("cards")}
-                  >
-                    <svg
-                      xmlns="http://www.w3.org/2000/svg"
-                      width="16"
-                      height="16"
-                      fill="currentColor"
-                      viewBox="0 0 16 16"
-                    >
-                      <path d="M0 4a2 2 0 0 1 2-2h12a2 2 0 0 1 2 2v8a2 2 0 0 1-2 2H2a2 2 0 0 1-2-2V4zm2-1a1 1 0 0 0-1 1v8a1 1 0 0 0 1 1h12a1 1 0 0 0 1-1V4a1 1 0 0 0-1-1H2z" />
-                      <path d="M4 5.5a.5.5 0 0 1 .5-.5h7a.5.5 0 0 1 .5.5v2a.5.5 0 0 1-.5.5h-7a.5.5 0 0 1-.5-.5v-2z" />
-                    </svg>
-                    Tarjetas
-                  </button>
+            {/* Controles de Visualización */}
+            <ViewControls 
+              viewMode={viewMode}
+              toggleViewMode={toggleViewMode}
+              isMobile={isMobile}
+              columnsDropdownOpen={columnsDropdownOpen}
+              setColumnsDropdownOpen={setColumnsDropdownOpen}
+              activeTab={activeTab}
+              visibleColumns={visibleColumns}
+              toggleColumn={toggleColumn}
+              columnToggleRef={columnToggleRef}
+              itemsPerPage={itemsPerPage}
+              handleItemsPerPageChange={handleItemsPerPageChange}
+            />
 
-                  {viewMode === "table" && !isMobile && (
-                    <div className="relative z-30" ref={columnToggleRef}>
-                      <ColumnSelector
-                        activeTab={activeTab}
-                        columnsDropdownOpen={columnsDropdownOpen}
-                        setColumnsDropdownOpen={setColumnsDropdownOpen}
-                        visibleColumns={visibleColumns}
-                        toggleColumn={toggleColumn}
-                        columnToggleRef={columnToggleRef}
-                      />
-                    </div>
-                  )}
-                </div>
-
-                {/* Selector de elementos por página */}
-                <div className="flex items-center">
-                  <label className="text-sm text-gray-400 mr-2">Mostrar:</label>
-                  <select
-                    className="bg-gray-800 border border-gray-700 text-gray-300 rounded-md py-1 px-2 text-sm cursor-pointer transition-colors duration-200 hover:border-blue-500/30"
-                    value={itemsPerPage}
-                    onChange={(e) => {
-                      handleItemsPerPageChange(Number(e.target.value));
-                    }}
-                  >
-                    {[10, 15, 25, 50].map((value) => (
-                      <option key={value} value={value}>
-                        {value}
-                      </option>
-                    ))}
-                  </select>
-                  <span className="text-sm text-gray-400 ml-2">elementos</span>
-                </div>
-              </div>
-            </div>
-
-            <div ref={contentRef} className="relative min-h-[200px]">
-              <div className="view-transition">
-                {viewMode === "table" && !isMobile ? (
-                  <TableComponent
-                    usuarios={activeTab === "usuarios" ? items || [] : []}
-                    investigadores={
-                      activeTab === "investigadores" ? items || [] : []
-                    }
-                    proyectos={activeTab === "proyectos" ? items || [] : []}
-                    visibleColumns={columns || []}
-                    onCopy={copyToClipboard}
-                    onEdit={onEdit}
-                    onDelete={onDelete}
-                  />
-                ) : (
-                  <div className="grid grid-cols-1 sm:grid-cols-2 lg:grid-cols-3 gap-4">
-                    <CardComponent
-                      items={items || []}
-                      onCopy={copyToClipboard}
-                      onEdit={onEdit}
-                      onDelete={onDelete}
-                    />
-                  </div>
-                )}
-
-                {/* Paginación - Con mejor espaciado */}
-                <div className="mt-8 pt-4 border-t border-gray-700/50 flex justify-center">
-                  <Pagination
-                    totalItems={totalItems}
-                    currentPage={currentPage}
-                    itemsPerPage={itemsPerPage}
-                    paginate={handlePageChange}
-                  />
-                </div>
-              </div>
-            </div>
+            {/* Contenido */}
+            <ContentDisplay 
+              contentRef={contentRef}
+              viewMode={viewMode}
+              isMobile={isMobile}
+              TableComponent={TableComponent}
+              CardComponent={CardComponent}
+              activeTab={activeTab}
+              items={items}
+              columns={columns}
+              onCopy={copyToClipboard}
+              onEdit={onEdit}
+              onDelete={onDelete}
+              totalItems={totalItems}
+              currentPage={currentPage}
+              itemsPerPage={itemsPerPage}
+              handlePageChange={handlePageChange}
+            />
           </div>
         </div>
       )}
